app/requirements.txt
prometheus-client==0.16.0
python-dateutil==2.8.2